feat : club API 파일 통일
